Restructure Gulp tasks to ensure correct tasks are loaded for correct environment
