+++ gulp/tasks/dev/lint-css.js
@@ -1,5 +1,5 @@
 const scsslint = require('gulp-scss-lint');
-const paths = require('../paths');
+const paths = require('../../paths');
 
 module.exports = (gulp) => {
   return gulp.src(`${paths.sourceStyles}/**/*.s+(a|c)ss`)
