--- gulp/tasks/lint.js
@@ -1,3 +1,0 @@
-module.exports = {
-  dep: ['lint-js', 'lint-css'],
-};
