+++ gulp/tasks/shared/clean.js
