+++ gulp/tasks/shared/fonts.js
