+++ gulp/tasks/shared/images.js
