+++ gulp/tasks/shared/rev.js
